requirements.txt
python-docx>=0.8.11
pdfplumber>=0.9.0
openpyxl>=3.1.2
openai>=1.0.0
python-dotenv>=1.0.0
